refactor(st-09015): split multi-agent node responsibilities
