+++ packages/patterns/src/multi-agent/nodes/shared.ts
@@ -1,0 +1,82 @@
+import { HumanMessage, SystemMessage } from '@langchain/core/messages';
+import { toLangChainTools, type Tool } from '@agentforge/core';
+import { createPatternLogger } from '../../shared/deduplication.js';
+import type { MultiAgentStateType } from '../state.js';
+import type { AgentMessage, TaskAssignment } from '../schemas.js';
+import type { WorkerConfig } from '../types.js';
+
+export const logger = createPatternLogger('agentforge:patterns:multi-agent:nodes');
+
+export function createGeneratedId(prefix: string): string {
+  return `${prefix}-${Date.now()}-${Math.random().toString(36).slice(2, 11)}`;
+}
+
+export function createTaskAssignmentId(): string {
+  return `task_${Date.now()}_${Math.random().toString(36).slice(2, 11)}`;
+}
+
+export function getLatestTaskContent(state: MultiAgentStateType): string {
+  return state.messages[state.messages.length - 1]?.content || state.input;
+}
+
+export function createTaskAssignments(workerIds: string[], task: string): TaskAssignment[] {
+  return workerIds.map((workerId) => ({
+    id: createTaskAssignmentId(),
+    workerId,
+    task,
+    priority: 5,
+    assignedAt: Date.now(),
+  }));
+}
+
+export function createAssignmentMessages(assignments: TaskAssignment[]): AgentMessage[] {
+  return assignments.map((assignment) => ({
+    id: createGeneratedId('msg'),
+    from: 'supervisor',
+    to: [assignment.workerId],
+    type: 'task_assignment',
+    content: assignment.task,
+    timestamp: Date.now(),
+    metadata: {
+      assignmentId: assignment.id,
+      priority: assignment.priority,
+    },
+  }));
+}
+
+export function findCurrentAssignment(
+  state: MultiAgentStateType,
+  workerId: string
+): TaskAssignment | undefined {
+  return state.activeAssignments.find(
+    (assignment) =>
+      assignment.workerId === workerId &&
+      !state.completedTasks.some((task) => task.assignmentId === assignment.id)
+  );
+}
+
+export function convertWorkerToolsForLangChain(tools: WorkerConfig['tools']) {
+  const safeTools = tools ?? [];
+  return toLangChainTools(
+    safeTools as unknown as Parameters<typeof toLangChainTools>[0]
+  );
+}
+
+export function serializeModelContent(content: unknown, fallback: string): string {
+  if (typeof content === 'string') {
+    return content;
+  }
+
+  try {
+    const serialized = JSON.stringify(content);
+    return typeof serialized === 'string' ? serialized : fallback;
+  } catch {
+    return fallback;
+  }
+}
+
+export function createWorkerMessages(systemPrompt: string, task: string) {
+  return [new SystemMessage(systemPrompt), new HumanMessage(task)];
+}
+
+export type WorkerTool = Tool<never, unknown>;
