+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -3,7 +3,7 @@
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
 **Last Updated:** 2026-03-25
-**Active Story:** ST-09015 (Ready)
+**Active Story:** ST-09015 (In Progress)
 
 ---
